requirements.txt
fastapi
uvicorn
SQLAlchemy
Jinja2
python-dotenv
mcp[sse]
python-multipart
